requirements.txt
# APEX Trader v2 — Python dependencies
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
ccxt>=4.2.0
yfinance>=0.2.36
alpaca-trade-api>=3.0.2
anthropic>=0.25.0
openai>=1.14.0
aiohttp>=3.9.0
